+++ initdb.d/81-create-exam-mode-table.sql
@@ -1,4 +1,3 @@
 DROP TABLE IF EXISTS exam_mode;
 CREATE TABLE exam_mode (
   exam BOOLEAN DEFAULT FALSE);
-