+++ src/typing_ex/view/page.clj
@@ -2,14 +2,15 @@
   (:refer-clojure :exclude [abs])
   (:require
    [ataraxy.response :as response]
+   #_[clojure.string :as str]
    [hiccup.page :refer [html5]]
    [hiccup.form :refer [form-to text-field password-field submit-button]]
    [java-time]
    [ring.util.anti-forgery :refer [anti-forgery-field]]
-   #_[taoensso.timbre :as timbre]
-   [typing-ex.plot :refer [scatter]]))
+   [typing-ex.plot :refer [scatter]]
+   [clojure.test :as t]))
 
-(def ^:private version "1.15.13")
+(def ^:private version "1.18.4")
 
 (defn page [& contents]
   [::response/ok
@@ -19,8 +20,8 @@
      [:meta {:name "viewport" :content "width=device-width, initial-scale=1"}]]
     [:link
      {:rel "stylesheet"
-      :href "https://stackpath.bootstrapcdn.com/bootstrap/4.5.0/css/bootstrap.min.css"
-      :integrity "sha384-9aIt2nRpC12Uk9gS9baDl411NQApFmC26EwAOH8WgZl5MYYxFfc+NcPb1dKGj7Sk"
+      :href "https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/css/bootstrap.min.css"
+      :integrity "sha384-rbsA2VBKQhggwzxH7pPCaAqO46MgnOM80zW1RWuH61DGLwZJEdK2Kadq2F9CUG65"
       :crossorigin "anonymous"}]
     [:link
      {:rel "stylesheet"
@@ -44,82 +45,118 @@
     (submit-button  "login"))
    [:br]
    [:ul
-    [:li "焦らず、ゆっくり、正確にタイピングが上達の早道。"]
+    [:li "焦らず、ゆっくり、正しい指使いがタイピングが上達の早道。"]
     [:li "10 分練習したら休憩入れよう。"]
     [:li "練習しないと平常点にならない。"]]))
-
-;; right place, here?
-(defn- count-ex-days [days login]
-  (->> days
-       (filter #(= (:login %) login))
-       count))
 
 (defn- headline
   "scores-page の上下から呼ぶ。ボタンの並び。他ページで使ってもよい。"
   [n]
   [:div {:style "margin-left:1rem;"}
     [:div.row
-     [:div.d-inline
+     [:div.d-inline-
       [:a {:href "/" :class "btn btn-primary btn-sm"} "Go!"]
-      " "
-      #_[:a {:href "/sum/7" :class "btn btn-primary btn-sm"} "D.P."]]
-     "&nbsp;|&nbsp;"
-     [:div.d-inline
-      (form-to
-       [:get "/recent"]
-       (text-field {:size 2
-                    :value n
-                    :style "text-align:right"}
-                   "n")
-       "days "
-       (submit-button {:class "btn btn-primary btn-sm"
-                       :name "total"}
-                      "total")
-       " "
-       (submit-button {:class "btn btn-primary btn-sm"
-                       :name "max"}
-                      "max"))]
-     "&nbsp;|&nbsp;"
-     [:div.d-inline
-      [:a {:href "/daily" :class "btn btn-danger btn-sm"}
-       "Today"]
-      " "
+      "&nbsp;"
+      [:a {:href "/rc" :class "btn btn-info btn-sm"} "RC"]
+      "&nbsp;"
       [:a {:href "https://wil.melt.kyutech.ac.jp/"
            :class "btn btn-info btn-sm"}
        "WIL"]
-      " "
+      "&nbsp;"
       [:a {:href "http://qa.melt.kyutech.ac.jp/"
            :class "btn btn-info btn-sm"}
        "QA"]
-      " "
+      "&nbsp;"
       [:a {:href "http://mt.melt.kyutech.ac.jp/"
            :class "btn btn-info btn-sm"}
        "MT"]
-      " "
+      "&nbsp;"
       [:a {:href "http://l22.melt.kyutech.ac.jp/"
            :class "btn btn-info btn-sm"}
        "L22"]
-      " "
-      [:a {:href "/logout" :class "btn btn-warning btn-sm"} "logout"]]]])
+      "&nbsp;"
+      [:a {:href "/logout" :class "btn btn-warning btn-sm"} "Logout"]]]
+     [:div.row
+      [:div.d-inline-flex
+       [:a {:href "/todays" :class "btn btn-danger btn-sm"}
+        "todays"]
+       "&nbsp;"
+       (form-to
+        [:get "/recent"]
+        (text-field {:size 2
+                     :value n
+                     :style "text-align:right"}
+                    "n")
+        " days → "
+        (submit-button {:class "btn btn-primary btn-sm"
+                        :name "kind"}
+                       "total")
+        "&nbsp;"
+        (submit-button {:class "btn btn-primary btn-sm"
+                        :name "kind"}
+                       "days")
+        "&nbsp;"
+        (submit-button {:class "btn btn-primary btn-sm"
+                        :name "kind"}
+                       "max"))]]])
 
-(defn scores-page [max-pt ex-days user days]
+(defn scores-page
+  "maxpt: 最高点
+   ex-days: 練習日数
+   user: アカウント
+   days: 何日間のデータか？"
+  [max-pt _ex-days user days]
   (page
    [:h2 "Typing: Last " days " days Maxes"]
    (headline days)
-   #_[:p "直近 " days " 日間のユーザ毎最高得点。カッコは通算練習日数。<br>
-情報リテラシー以外の科目も大切にしよう。"]
-   [:p "スコア伸びないのは練習足りない？ 情報リテラシー以外の科目も大切に。"]
-   (into [:ol
-          (for [{:keys [max login]} max-pt]
-            [:li
-             max
-             (format "(%d) " (count-ex-days ex-days login))
-             [:a {:href (str "/record/" login)
-                  :class (if (= login user) "yes" "other")}
-              login]])])
+   [:div {:style "margin-left:1rem;"}
+    [:p "瞬間最大風速。" [:br]
+         "(タイプの正確さ) + (あまりの秒数)なので、理論的な最高得点は 159。"]
+    (into [:ol
+           (for [{:keys [max login]} max-pt]
+             [:li
+              max
+              " "
+              [:a {:href (str "/record/" login)
+                   :class (if (= login user) "yes" "other")}
+               login]])])]
    (headline days)))
 
-;; not good
+(defn- count-ex-days
+  [days login]
+  ;; 引数 days の中身は、[{:login ... :date ...} ...]
+  ;; (println "days:" (str days))
+  (->> days
+       (filter #(= (:login %) login))
+       count))
+
+(defn ex-days-page
+  "ex-days: 練習日数
+   user: アカウント
+   days: 何日間のデータか？"
+  [ex-days user days]
+  (let [logins (->> ex-days (map :login) distinct)
+        data (->> (for [login logins]
+                    [(count-ex-days ex-days login) login])
+                  (sort-by first)
+                  reverse)]
+    (page
+     [:h2 "Typing: Last " days " days Maxes"]
+     (headline days)
+     [:div {:style "margin-left:1rem;"}
+      [:p "「1 日 10 回未満はノーカウント」て言うと 10 回で終わる人いるからなあ。" [:br]
+       "そういうスタンスじゃ上手にならんやろ。"]
+      (into [:ol
+             (for [[count login] data]
+               [:li
+                (format "(%d) " count)
+                " "
+                [:a {:href (str "/record/" login)
+                     :class (if (= login user) "yes" "other")}
+                 login]])])]
+     (headline days))))
+
+;; FIXME
 (defn- ss
   "shorten string"
   [s]
@@ -139,19 +176,32 @@
               (map str)
               (group-by identity))))
 
-;; 平均を求めるのに、DB 引かなくても ret から求めればいい。
-;; ret は lazySeq
-;; 1.5.8 Exercise days
-(defn svg-self-records [login ret me? admin?]
+;; ret is a  lazySeq. should use mapv?
+(defn svg-self-records
+  [login ret _me? _admin?]
   (let [positives (map #(assoc % :pt (max 0 (:pt %))) ret)
         avg (/ (reduce + (map :pt (take 10 (reverse positives)))) 10.0)
-        todays (filter #(today? (:timestamp %)) ret)]
+        todays (filter #(today? (:timestamp %)) ret)
+        ;; _ (def p positives)
+        ;; _ (def t todays)
+        ]
     (page
-     [:h2 "Typing: " login " records"]
+     [:h2 "Typing: " login " Records"]
      [:p "付け焼き刃はもろい。毎日 10 分 x 3 セット。"]
-     [:div (scatter 300 150 positives)]
+     [:div.d-inline-flex
+      [:div.px-2.mx-auto
+       (scatter 300 150 (map :pt positives))
+       [:br]
+       [:b "TOTAL"]]
+      (when (< 9 (count todays))
+        [:div.px-2.mx-auto
+         (scatter 300 150 (map :pt todays))
+         [:br]
+         [:b "TODAYS"]])
+      [:div.px-2]]
      [:br]
-     (when (or me? admin?)
+     [:br]
+     (when true ;; (or me? admin?)
        [:ul
         [:li "Max " (apply max (map :pt positives))]
         [:li "Average (last 10) " avg]
@@ -160,6 +210,12 @@
         [:li "Last Exercise " (ss (str (:timestamp (last ret))))]])
      [:p [:a {:href "/" :class "btn btn-primary btn-sm"} "Go!"]])))
 
+;; (comment
+;;   p
+;;   t
+;;   :rcf)
+
+;; using?
 (defn active-users-page [ret]
   (page
    [:h2 "Typing: Last 40 trials"]
@@ -169,29 +225,60 @@
          (for [[u & _] ret]
            [:li (ss (:timestamp u)) " " (:login u)]))))
 
+;; view of /todays
 (defn todays-act-page [ret login]
+  ;;(println "todays-act-page " (str ret))
   (page
-   [:h2 "Typing: todays"]
-   [:p "本日の Typing ユーザ。重複を省いて最終利用時間で並べ替え。"]
-   (into [:ol]
-         (for [r ret]
-           [:li (ss (java-time/local-date-time (:timestamp r)))
-            " "
-            [:a {:href (str "/record/" (:login r))
-                 :class (if (= login (:login r)) "yes" "other")}
-             (:login r)]]))))
+   [:h2 "Typing: Todays"]
+   (headline 7)
+   [:div {:style "margin-left:1rem;"}
+    [:p "平常点が必要な人は見せかけじゃなく、実質的に平常から回数を重ねないと。" [:br]
+     "練習したフリはタメにならない。"]
+    (into [:ol]
+          (for [r ret]
+            [:li (ss (java-time/local-date-time (:timestamp r)))
+             " "
+             [:a {:href (str "/record/" (:login r))
+                  :class (if (= login (:login r)) "yes" "other")}
+              (:login r)]]))]
+   (headline 7)))
 
 (defn sums-page [ret user n]
   (page
-   [:h2 "Typing: Daily Points"]
+   [:h2 "Typing: Last " n " days Totals"]
    (headline n)
-   [:p "7 days で 3000 点が合格ラインだったのは先週。今週は 3500 くらいか。"]
-   (into [:ol]
-         (for [r ret]
-           (let [login (:login r)]
-             [:li (:sum r)
-              " "
-              [:a {:href (str "/record/" login)
-                   :class (if (= user login) "yes" "other")}
-               login]])))
+   [:div {:style "margin-left:1rem;"}
+    [:p "毎日ちょっとずつ点数アップが一番。一度にたくさんやっても身につかないよ。"]
+    (into [:ol]
+          (for [r ret]
+            (let [login (:login r)
+                  sum (:sum r)]
+              (when (< -1 sum)
+                [:li sum
+                 " "
+                 [:a {:href (str "/record/" login)
+                      :class (if (= user login) "yes" "other")}
+                  login]]))))]
    (headline n)))
+
+(defn stat-page [stat]
+  (page
+   [:h2 "Typing: Stat"]
+   (form-to
+    [:post "/stat"]
+    (anti-forgery-field)
+    [:input {:name "stat" :placeholder stat}]
+    [:p "normal, roll-call, exam"]
+    [:input {:type "submit"}])))
+
+;; roll-call
+;; FIXME: 表示で工夫するよりも、データベースに入れる時に加工するか？
+(defn rc-page [ret]
+  (page
+   [:h2 "Typing: 出席データ(" (-> ret first :login) ")"]
+   [:p "タイピングの背景が黄色い間にタイプ練習終了した時刻を記録している。"
+    "タイピングのバージョンが 1.16.7 より低い時は"
+    "2 週目でやった「閲覧履歴の消去」でバージョンアップしよう。"]
+   [:ul {:class "roll-call"}
+    (for [r ret]
+      [:li (str (:created_at r))])]))